+++ pcs_test/tier0/lib/booth/test_cib.py
@@ -2,10 +2,10 @@ from unittest import TestCase
 
 from lxml import etree
 
-from pcs.lib.booth.cib import get_ticket_names
+from pcs.lib.booth import cib as cib_commands
 
 
-class GetTicketNames(TestCase):
+class GetTicketNamesBase:
     def setUp(self):
         self.cib = etree.fromstring(
             """
@@ -23,11 +23,29 @@ class GetTicketNames(TestCase):
         """
         )
 
+
+class GetBoothTicketNames(GetTicketNamesBase, TestCase):
     def test_default(self):
-        self.assertEqual(["T1", "T2"], get_ticket_names(self.cib))
+        self.assertEqual(
+            ["T1", "T2"], cib_commands.get_booth_ticket_names(self.cib)
+        )
 
     def test_custom_booth_name(self):
-        self.assertEqual(["T3"], get_ticket_names(self.cib, "custom_booth"))
+        self.assertEqual(
+            ["T3"],
+            cib_commands.get_booth_ticket_names(self.cib, "custom_booth"),
+        )
 
     def test_no_tickets(self):
-        self.assertEqual([], get_ticket_names(self.cib, "nonexistent_booth"))
+        self.assertEqual(
+            [],
+            cib_commands.get_booth_ticket_names(self.cib, "nonexistent_booth"),
+        )
+
+
+class GetTicketNames(GetTicketNamesBase, TestCase):
+    def test_success(self):
+        self.assertEqual(
+            ["T1", "T2", "T3", "T-self-managed"],
+            cib_commands.get_ticket_names(self.cib),
+        )