+++ pcs_test/tier1/test_booth.py
@@ -709,3 +709,33 @@ class TicketCleanup(BoothMixinNoFiles, TestCase):
                 "\nUsage: pcs booth <command>\n    ticket cleanup [ticket]\n"
             ),
         )
+
+
+class BoothTicketOperationBase(BoothMixinNoFiles):
+    command = ""
+
+    def test_not_enough_args(self):
+        self.assert_pcs_fail(
+            ["booth", "ticket", self.command],
+            stderr_start=(
+                "\nUsage: pcs booth <command>\n"
+                f"    ticket {self.command} <ticket>\n"
+            ),
+        )
+
+    def test_too_many_args(self):
+        self.assert_pcs_fail(
+            ["booth", "ticket", self.command, "a", "b"],
+            stderr_start=(
+                "\nUsage: pcs booth <command>\n"
+                f"    ticket {self.command} <ticket>\n"
+            ),
+        )
+
+
+class TicketUnstandby(BoothTicketOperationBase, TestCase):
+    command = "unstandby"
+
+
+class TicketStandby(BoothTicketOperationBase, TestCase):
+    command = "standby"