+++ CHANGELOG.md
@@ -5,11 +5,14 @@
 ### Added
 - Command `pcs booth ticket cleanup` that enables removal of booth tickets
   still loaded in CIB after their removal from booth configuration ([RHEL-7602])
+- Commands `pcs booth ticket standby` and `pcs booth ticket unstandby` which
+  allow for managing the state of the ticket ([RHEL-12709])
 
 ### Fixed
 - Specify the meaning of zero value timeout in `pcs status wait` ([RHEL-44719])
 
 [RHEL-7602]: https://issues.redhat.com/browse/RHEL-7602
+[RHEL-12709]: https://issues.redhat.com/browse/RHEL-12709
 [RHEL-44719]: https://issues.redhat.com/browse/RHEL-44719